merge 2.0.0 platform-sh-starter

+++ web/wp-content/plugins/gravityforms/includes/api.php
@@ -42,11 +42,16 @@ class GFAPI {
 			return false;
 		}
 
+		$form_info = GFFormsModel::get_form( $form_id, true );
+		if ( ! $form_info ) {
+			return false;
+		}
+
 		// Loading form columns into meta.
-		$form_info            = GFFormsModel::get_form( $form_id, true );
 		$form['is_active']    = $form_info->is_active;
 		$form['date_created'] = $form_info->date_created;
 		$form['is_trash']     = $form_info->is_trash;
+		$form['title']        = $form_info->title;
 
 		return $form;
 
@@ -56,19 +61,22 @@ class GFAPI {
 	 * Returns all the form objects.
 	 *
 	 * @since  1.8.11.5
+	 * @since  2.5 added $sort_column and $sort_dir parameters.
 	 * @access public
 	 *
 	 * @uses GFFormsModel::get_form_ids()
 	 * @uses GFAPI::get_form()
 	 *
-	 * @param bool $active True if active forms are returned. False to get inactive forms. Defaults to true.
-	 * @param bool $trash  True if trashed forms are returned. False to exclude trash. Defaults to false.
+	 * @param bool   $active      True if active forms are returned. False to get inactive forms. Defaults to true.
+	 * @param bool   $trash       True if trashed forms are returned. False to exclude trash. Defaults to false.
+	 * @param string $sort_column The column to sort the results on.
+	 * @param string $sort_dir    The sort direction, ASC or DESC.
 	 *
 	 * @return array The array of Form Objects.
 	 */
-	public static function get_forms( $active = true, $trash = false ) {
+	public static function get_forms( $active = true, $trash = false, $sort_column = 'id', $sort_dir = 'ASC' ) {
 
-		$form_ids = GFFormsModel::get_form_ids( $active, $trash );
+		$form_ids = GFFormsModel::get_form_ids( $active, $trash, $sort_column, $sort_dir );
 		if ( empty( $form_ids ) ) {
 			return array();
 		}
@@ -218,6 +226,10 @@ class GFAPI {
 			return new WP_Error( 'missing_form_id', __( 'Missing form id', 'gravityforms' ) );
 		}
 
+		if ( isset( $form['title'] ) ) {
+			$form['title'] = self::unique_title( $form['title'], $form_id );
+		}
+
 		if ( isset( $form['fields'] ) ) {
 
 			// Make sure the formId is correct.
@@ -298,9 +310,6 @@ class GFAPI {
 	 * @since  1.8.3.15
 	 * @access public
 	 *
-	 * @uses GFFormsModel::get_form_table_name()
-	 * @uses GFFormsModel::get_form_db_columns()
-	 *
 	 * @param array  $form_ids     The IDs of the forms to update.
 	 * @param string $property_key The name of the column in the database e.g. is_trash, is_active, title.
 	 * @param mixed  $value        The new value.
@@ -314,12 +323,27 @@ class GFAPI {
 			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
 		}
 
-		$table        = GFFormsModel::get_form_table_name();
-
+		$table      = GFFormsModel::get_form_table_name();
 		$db_columns = GFFormsModel::get_form_db_columns();
 
 		if ( ! in_array( strtolower( $property_key ), $db_columns ) ) {
 			return new WP_Error( 'property_key_incorrect', __( 'Property key incorrect', 'gravityforms' ) );
+		}
+
+		if ( 'title' == $property_key ) {
+			if ( count( $form_ids ) === 1 ) {
+				$value = self::unique_title( $value, $form_ids[0] );
+			} else {
+				foreach ( $form_ids as $form_id ) {
+					$result = self::update_forms_property( array( $form_id ), $property_key, $value );
+					if ( is_wp_error( $result ) ) {
+						// If the result is an error, return the error right away.
+						return $result;
+					}
+				}
+
+				return $result;
+			}
 		}
 
 		$value = esc_sql( $value );
@@ -327,7 +351,7 @@ class GFAPI {
 			$value = sprintf( "'%s'", $value );
 		}
 		$in_str_arr = array_fill( 0, count( $form_ids ), '%d' );
-		$in_str     = join( $in_str_arr, ',' );
+		$in_str     = implode( ',', $in_str_arr );
 		$result     = $wpdb->query(
 			$wpdb->prepare(
 				"
@@ -346,8 +370,6 @@ class GFAPI {
 	 *
 	 * @since  1.8.3.15
 	 * @access public
-	 *
-	 * @uses GFAPI::update_forms_property()
 	 *
 	 * @param array|int $form_id      The ID of the forms to update.
 	 * @param string    $property_key The name of the column in the database e.g. is_trash, is_active, title.
@@ -403,11 +425,6 @@ class GFAPI {
 	 * @access public
 	 * @global $wpdb
 	 *
-	 * @uses GFFormsModel::is_unique_title()
-	 * @uses GFFormsModel::insert_form()
-	 * @uses GFAPI::set_property_as_key()
-	 * @uses GFFormsModel::update_form_meta()
-	 *
 	 * @param array $form_meta The Form object.
 	 *
 	 * @return int|WP_Error Either the new Form ID or a WP_Error instance.
@@ -432,12 +449,7 @@ class GFAPI {
 		}
 
 		// Making sure title is not duplicate.
-		$title = $form_meta['title'];
-		$count = 2;
-		while ( ! RGFormsModel::is_unique_title( $title ) ) {
-			$title = $form_meta['title'] . "($count)";
-			$count ++;
-		}
+		$title = self::unique_title( $form_meta['title'] );
 
 		// Inserting form.
 		$form_id = RGFormsModel::insert_form( $title );
@@ -448,23 +460,16 @@ class GFAPI {
 		// Updating object's id property.
 		$form_meta['id'] = $form_id;
 
+		// Adding markup version. Increment this when we make breaking changes to form markup.
+		$form_meta['markupVersion'] = rgar( $form_meta, 'markupVersion' ) ? $form_meta['markupVersion'] : 2;
+
 		// Add default confirmation if form has no confirmations.
 		if ( ! isset( $form_meta['confirmations'] ) || empty( $form_meta['confirmations'] ) ) {
 
-			// Generate confirmation ID.
-			$confirmation_id = uniqid();
+			$confirmation = GFFormsModel::get_default_confirmation();
 
 			// Add default confirmation to form.
-			$form_meta['confirmations'][ $confirmation_id ] = array(
-				'id'          => $confirmation_id,
-				'name'        => __( 'Default Confirmation', 'gravityforms' ),
-				'isDefault'   => true,
-				'type'        => 'message',
-				'message'     => __( 'Thanks for contacting us! We will get in touch with you shortly.', 'gravityforms' ),
-				'url'         => '',
-				'pageId'      => '',
-				'queryString' => '',
-			);
+			$form_meta['confirmations'] = array( $confirmation['id'] => $confirmation );
 
 		}
 
@@ -885,10 +890,8 @@ class GFAPI {
 		}
 		$source_url = $entry['source_url'];
 
-		if ( ! isset( $entry['user_agent'] ) ) {
-			$entry['user_agent'] = 'API';
-		}
-		$user_agent = $entry['user_agent'];
+		$entry['user_agent'] = isset( $entry['user_agent'] ) ? sanitize_text_field( $entry['user_agent'] ) : 'API';
+		$user_agent          = $entry['user_agent'];
 
 		if ( empty( $entry['currency'] ) ) {
 			$entry['currency'] = GFCommon::get_currency();
@@ -996,8 +999,7 @@ class GFAPI {
 
 		foreach ( $form['fields'] as $field ) {
 			/* @var GF_Field $field */
-			$type = GFFormsModel::get_input_type( $field );
-			if ( in_array( $type, array( 'html', 'page', 'section' ) ) ) {
+			if ( $field->displayOnly ) {
 				continue;
 			}
 
@@ -1218,7 +1220,7 @@ class GFAPI {
 		$request_ip     = rgars( $form, 'personalData/preventIP' ) ? '' : GFFormsModel::get_ip();
 		$ip             = isset( $entry['ip'] ) ? $entry['ip'] : $request_ip;
 		$source_url     = isset( $entry['source_url'] ) ? $entry['source_url'] : esc_url_raw( GFFormsModel::get_current_page_url() );
-		$user_agent     = isset( $entry['user_agent'] ) ? $entry['user_agent'] : 'API';
+		$user_agent     = isset( $entry['user_agent'] ) ? sanitize_text_field( $entry['user_agent'] ) : 'API';
 		$currency       = isset( $entry['currency'] ) ? $entry['currency'] : GFCommon::get_currency();
 		$payment_status = isset( $entry['payment_status'] ) ? sprintf( "'%s'", esc_sql( $entry['payment_status'] ) ) : 'NULL';
 		$payment_date   = strtotime( rgar( $entry, 'payment_date' ) ) ? sprintf( "'%s'", gmdate( 'Y-m-d H:i:s', strtotime( "{$entry['payment_date']}" ) ) ) : 'NULL';
@@ -1259,7 +1261,7 @@ class GFAPI {
 
 		foreach ( $form['fields'] as $field ) {
 			/* @var GF_Field $field */
-			if ( in_array( $field->type, array( 'html', 'page', 'section' ) ) ) {
+			if ( $field->displayOnly ) {
 				continue;
 			}
 			self::queue_batch_field_operation( $form, $entry, $field );
@@ -1418,6 +1420,173 @@ class GFAPI {
 				$result = GFFormsModel::update_entry_field_value( $form, $entry, $field, $lead_detail_id, $input_id, $value, $item_index );
 			}
 		}
+
+		return $result;
+	}
+
+	// ENTRY NOTES ------------------------------------------------
+
+	/**
+	 * Get notes based on search criteria.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param array $search_criteria Array of search criteria.
+	 * @param array $sorting Sort key and direction.
+	 * @return array|bool
+	 */
+	public static function get_notes( $search_criteria = array(), $sorting = null ) {
+
+		if ( ! $sorting ) {
+			$sorting = array(
+				'key'        => 'id',
+				'direction'  => 'ASC',
+				'is_numeric' => true,
+			);
+		}
+
+		$notes = GFFormsModel::get_notes( $search_criteria, $sorting );
+
+		if ( empty( $notes ) ) {
+			return false;
+		}
+
+		return $notes;
+	}
+
+	/**
+	 * Get note by ID.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int $note_id ID of the note to retrieve.
+	 * @return array|WP_Error
+	 */
+	public static function get_note( $note_id ) {
+		$note = GFFormsModel::get_notes( array( 'id' => $note_id ) );
+
+		if ( empty( $note ) ) {
+			return new WP_Error( 'note_not_found', __( 'Note not found', 'gravityforms' ) );
+		}
+
+		return $note[0];
+	}
+
+	/**
+	 * Create one note for an entry.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int    $entry_id ID of the entry to add the note to.
+	 * @param int    $user_id ID of the user to associate with the note.
+	 * @param string $user_name Name of the user to associate with the note.
+	 * @param string $note Text of the note.
+	 * @param string $note_type Note type.
+	 * @param null   $sub_type Not sub-type.
+	 * @return array|int|void|WP_Error
+	 */
+	public static function add_note( $entry_id, $user_id, $user_name, $note, $note_type = 'user', $sub_type = null ) {
+		if ( gf_upgrade()->get_submissions_block() ) {
+			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
+		}
+
+		if ( ! self::entry_exists( $entry_id ) ) {
+			return new WP_Error( 'invalid_entry', __( 'Invalid entry', 'gravityforms' ), $entry_id );
+		}
+
+		if ( empty( $note ) || ! is_string( $note ) ) {
+			return new WP_Error( 'invalid_note', __( 'Invalid or empty note', 'gravityforms' ), $entry_id );
+		}
+
+		$new_note = GFFormsModel::add_note( intval( $entry_id ), $user_id, $user_name, wp_kses_post( $note ), sanitize_text_field( $note_type ), sanitize_text_field( $sub_type ) );
+
+		return $new_note;
+	}
+
+	/**
+	 * Delete one note.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param int $note_id ID of the note to delete.
+	 * @return int|WP_Error ID of the deleted note.
+	 */
+	public static function delete_note( $note_id ) {
+		$result = GFFormsModel::delete_note( $note_id );
+
+		if ( ! $result ) {
+			return new WP_Error( 'invalid_note', __( 'Invalid note', 'gravityforms' ), $note_id );
+		}
+
+		return $result;
+	}
+
+	/**
+	 * Update a note.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param array $note {
+	 * 		Note data to update.
+	 *
+	 *		@type int    $entry_id     ID of the entry associated with the note.
+	 *		@type int    $user_id      ID of the user associated with the note.
+	 * 		@type string $user_name    Name of the user associated with the note.
+	 *		@type string $date_created Date and time the note was created, in SQL datetime format.
+	 *		@type string $value        The text of the note.
+	 *		@type string $note_type    The note type.
+	 *		@type string $sub_type     The note subtype.
+	 * }
+	 * @param int   $note_id ID of the note to update.
+	 * @return bool|WP_Error
+	 */
+	public static function update_note( $note, $note_id = '' ) {
+		if ( gf_upgrade()->get_submissions_block() ) {
+			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
+		}
+
+		if ( ! is_array( $note ) || empty( $note ) ) {
+			return new WP_Error( 'invalid_note_format', __( 'Invalid note format', 'gravityforms' ) );
+		}
+
+		if ( empty( $note_id ) ) {
+			if ( rgar( $note, 'id' ) ) {
+				$note_id = absint( $note['id'] );
+			}
+		} else {
+			$note_id = absint( $note_id );
+		}
+
+		if ( empty( $note_id ) ) {
+			return new WP_Error( 'missing_note_id', __( 'Missing note id', 'gravityforms' ) );
+		}
+
+		// make sure the note exists.
+		$current_note = self::get_note( $note_id );
+		if ( ! $current_note || is_wp_error( $current_note ) ) {
+			return new WP_Error( 'note_not_found', __( 'Note not found', 'gravityforms' ) );
+		}
+
+		$note_properties = array(
+			'id',
+			'entry_id',
+			'user_id',
+			'user_name',
+			'date_created',
+			'value',
+			'note_type',
+			'sub_type',
+		);
+
+		$current_note_array = (array) $current_note;
+
+		foreach ( $note_properties as $property ) {
+			if ( ! isset( $note[ $property ] ) ) {
+				$note[ $property ] = $current_note_array[ $property ];
+			}
+		}
+
+		$result = GFFormsModel::update_note( $note['id'], $note['entry_id'], $note['user_id'], $note['user_name'], $note['date_created'], $note['value'], $note['note_type'], $note['sub_type'] );
 
 		return $result;
 	}
@@ -1606,19 +1775,26 @@ class GFAPI {
 	 * @access public
 	 * @global $wpdb
 	 *
-	 * @param mixed  $feed_ids   The ID of the Feed or an array of Feed IDs.
-	 * @param int    $form_id    The ID of the Form to which the Feeds belong.
-	 * @param string $addon_slug The slug of the add-on to which the Feeds belong.
-	 * @param bool   $is_active  If the feed is active.
+	 * @param mixed       $feed_ids   The ID of the Feed or an array of Feed IDs.
+	 * @param null|int    $form_id    The ID of the Form to which the Feeds belong.
+	 * @param null|string $addon_slug The slug of the add-on to which the Feeds belong.
+	 * @param bool|null   $is_active  Indicates if only active or inactive feeds should be returned. Use null to return both.
 	 *
 	 * @return array|WP_Error Either an array of Feed objects or a WP_Error instance.
 	 */
 	public static function get_feeds( $feed_ids = null, $form_id = null, $addon_slug = null, $is_active = true ) {
 		global $wpdb;
 
-		$table       = $wpdb->prefix . 'gf_addon_feed';
-		$where_arr   = array();
-		$where_arr[] = $wpdb->prepare( 'is_active=%d', $is_active );
+		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
+
+		$where_arr = array();
+		if ( null !== $is_active ) {
+			$where_arr[] = $wpdb->prepare( 'is_active=%d', $is_active );
+		}
 		if ( false === empty( $form_id ) ) {
 			$where_arr[] = $wpdb->prepare( 'form_id=%d', $form_id );
 		}
@@ -1627,17 +1803,19 @@ class GFAPI {
 		}
 		if ( false === empty( $feed_ids ) ) {
 			if ( ! is_array( $feed_ids ) ) {
-				$feed_ids = array( $feed_ids );
+				$where_arr[] = $wpdb->prepare( 'id=%d', $feed_ids );
+			} else {
+				$in_str_arr  = array_fill( 0, count( $feed_ids ), '%d' );
+				$in_str      = join( ',', $in_str_arr );
+				$where_arr[] = $wpdb->prepare( "id IN ($in_str)", $feed_ids );
 			}
-			$in_str_arr  = array_fill( 0, count( $feed_ids ), '%d' );
-			$in_str      = join( $in_str_arr, ',' );
-			$where_arr[] = $wpdb->prepare( "id IN ($in_str)", $feed_ids );
 		}
 
+		$sql = "SELECT * FROM {$table}";
 
-		$where = join( ' AND ', $where_arr );
-
-		$sql = "SELECT id, form_id, addon_slug, meta FROM {$table} WHERE $where";
+		if ( ! empty( $where_arr ) ) {
+			$sql .= ' WHERE ' . join( ' AND ', $where_arr );
+		}
 
 		$results = $wpdb->get_results( $sql, ARRAY_A );
 		if ( empty( $results ) ) {
@@ -1649,6 +1827,24 @@ class GFAPI {
 		}
 
 		return $results;
+	}
+
+	/**
+	 * Returns a specific feed.
+	 *
+	 * @since 2.4.24
+	 *
+	 * @param int $feed_id The ID of the feed to retrieve.
+	 *
+	 * @return array|WP_Error
+	 */
+	public static function get_feed( $feed_id ) {
+		$feeds = self::get_feeds( $feed_id, null, null, null );
+		if ( is_wp_error( $feeds ) ) {
+			return $feeds;
+		}
+
+		return $feeds[0];
 	}
 
 	/**
@@ -1670,6 +1866,10 @@ class GFAPI {
 		}
 
 		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
 
 		$sql = $wpdb->prepare( "DELETE FROM {$table} WHERE id=%d", $feed_id );
 
@@ -1744,7 +1944,12 @@ class GFAPI {
 			return new WP_Error( 'submissions_blocked', __( 'Submissions are currently blocked due to an upgrade in progress', 'gravityforms' ) );
 		}
 
-		$table          = $wpdb->prefix . 'gf_addon_feed';
+		$table = $wpdb->prefix . 'gf_addon_feed';
+
+		if ( ! GFCommon::table_exists( $table ) ) {
+			return self::get_missing_table_wp_error( $table );
+		}
+
 		$feed_meta_json = json_encode( $feed_meta );
 		$sql            = $wpdb->prepare( "INSERT INTO {$table} (form_id, meta, addon_slug) VALUES (%d, %s, %s)", $form_id, $feed_meta_json, $addon_slug );
 
@@ -1755,6 +1960,34 @@ class GFAPI {
 		}
 
 		return $wpdb->insert_id;
+	}
+
+	/**
+	 * Updates the specified feed with the given property value.
+	 *
+	 * @since 2.4.24
+	 *
+	 * @param int    $feed_id        The ID of the feed being updated.
+	 * @param string $property_name  The name of the property (column) being updated.
+	 * @param mixed  $property_value The new value of the specified property.
+	 *
+	 * @return bool|WP_Error
+	 */
+	public static function update_feed_property( $feed_id, $property_name, $property_value ) {
+		return GFFormsModel::update_feed_property( $feed_id, $property_name, $property_value );
+	}
+
+	/**
+	 * Returns the missing_table WP_Error.
+	 *
+	 * @since 2.4.22
+	 *
+	 * @param string $table The name of the table which does not exist.
+	 *
+	 * @return WP_Error
+	 */
+	private static function get_missing_table_wp_error( $table ) {
+		return new WP_Error( 'missing_table', sprintf( __( 'The %s table does not exist.', 'gravityforms' ), $table ) );
 	}
 
 	// NOTIFICATIONS ----------------------------------------------
@@ -1911,28 +2144,17 @@ class GFAPI {
 	/**
 	 * Checks whether a form ID exists.
 	 *
-	 * @since  1.8
+	 * @since 1.8
+	 * @since 2.4.24 Updated to use GFFormsModel::id_exists_in_table().
 	 */
 	public static function form_id_exists( $form_id ) {
-		global $wpdb;
-		$form_table_name = GFFormsModel::get_form_table_name();
-		$form_id         = intval( $form_id );
-		$result          = $wpdb->get_var(
-			$wpdb->prepare(
-				" SELECT count(id) FROM {$form_table_name}
-                  WHERE id=%d", $form_id
-			)
-		);
-
-		$result = intval( $result );
-
-		return $result > 0;
+		return GFFormsModel::id_exists_in_table( $form_id, GFFormsModel::get_form_table_name() );
 	}
 
 	/**
 	 * Checks if an entry exists for the supplied ID.
 	 *
-	 * @since 2.4.5.8
+	 * @since 2.4.6
 	 *
 	 * @param int $entry_id The ID to be checked.
 	 *
@@ -1940,6 +2162,19 @@ class GFAPI {
 	 */
 	public static function entry_exists( $entry_id ) {
 		return GFFormsModel::entry_exists( $entry_id );
+	}
+
+	/**
+	 * Checks if a feed exists for the supplied ID.
+	 *
+	 * @since 2.4.24
+	 *
+	 * @param int $feed_id The ID to be checked.
+	 *
+	 * @return bool
+	 */
+	public static function feed_exists( $feed_id ) {
+		return GFFormsModel::id_exists_in_table( $feed_id, GFFormsModel::get_addon_feed_table_name() );
 	}
 
 	/**
@@ -1968,6 +2203,20 @@ class GFAPI {
 			GFLogging::include_logger();
 			GFLogging::log_message( 'gravityformsapi', $message, KLogger::DEBUG );
 		}
+	}
+
+	/**
+	 * Make sure the form title is unique.
+	 *
+	 * @since 2.5
+	 *
+	 * @param string     $title
+	 * @param int|string $form_id
+	 *
+	 * @return string
+	 */
+	public static function unique_title( $title, $form_id = '' ) {
+		return GFFormsModel::maybe_increment_title( $title, $form_id );
 	}
 
 }
